requirement.txt
asgiref==3.8.1
Django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.15.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
